backend/requirements.txt
alembic==1.4.3
Authlib==0.14.3
fastapi==0.61.1
celery==5.0.0
redis==3.5.3
httpx==0.15.5
ipython==7.16.0
itsdangerous==1.1.0
Jinja2==2.11.2
psycopg2==2.8.6
pytest==6.1.0
requests==2.24.0
SQLAlchemy==1.3.19
starlette==0.13.6
uvicorn==0.12.1
passlib==1.7.2
bcrypt==3.2.0
sqlalchemy-utils==0.36.8
python-multipart==0.0.5
pyjwt==1.7.1
aiofiles
numpy==1.18.2
torch==1.6.0
torchvision==0.7.0
tensorboardX==2.0
matplotlib
tqdm==4.43.0
easydict==1.9
Pillow==7.1.2
scikit-image
opencv-contrib-python
pycocotools
moviepy